+++ src/ggrc-client/js/components/related-objects/proposals/related-proposals-item.js
@@ -12,7 +12,7 @@ import '../../diff/instance-mapping-fields-diff';
 import '../../diff/instance-list-fields-diff';
 import template from './templates/related-proposals-item.mustache';
 import {getPersonInfo} from '../../../plugins/ggrc_utils';
-import {getFormattedLocalDate} from '../../../plugins/utils/date-util';
+import {getFormattedLocalDate} from '../../../plugins/utils/date-utils';
 const tag = 'related-proposals-item';
 
 export default can.Component.extend({
