+++ src/ggrc-client/js/components/release-notes-menu-item/release-notes-menu-item.js
@@ -6,7 +6,7 @@
 import '../release-notes-modal/release-notes-modal';
 import template from './release-notes-menu-item.mustache';
 import {loadUserProfile, updateUserProfile} from '../../plugins/utils/user-utils';
-import {getFormattedUtcDate} from '../../plugins/utils/date-util';
+import {getFormattedUtcDate} from '../../plugins/utils/date-utils';
 
 const viewModel = can.Map.extend({
   define: {
