+++ src/ggrc-client/js/components/tests/datepicker_spec.js
@@ -4,7 +4,7 @@
  */
 
 import * as Utils from '../../plugins/ggrc_utils';
-import {DATE_FORMAT} from '../../plugins/utils/date-util';
+import {DATE_FORMAT} from '../../plugins/utils/date-utils';
 import {getComponentVM} from '../../../js_specs/spec_helpers';
 import Component from '../datepicker/datepicker';
